requirements.txt
# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Demo
gradio>=4.0.0

# Vision and Audio
openai-whisper>=20231117
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# LLM
ollama>=0.1.0
openai>=1.3.0

# MLops
mlflow>=2.8.0

# Utils
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
tqdm>=4.66.0

# Dev & Test
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
ruff>=0.4.0
pytest-cov>=4.1.0